feat: move sql utils to a separate package. Implemented a transaction preparer interface to add flexibility to the with transaction functionality
