+++ pkg/sqlutils/transaction_preparer.go
@@ -1,0 +1,8 @@
+package sqlutils
+
+import "github.com/jmoiron/sqlx"
+
+// TransactionPreparer is an interface used to initiate a sqlx.TX
+type TransactionPreparer interface {
+	Beginx() (*sqlx.Tx, error)
+}
