+++ pkg/sqlutils/transactions.go
@@ -1,4 +1,5 @@
-package storage
+// Package sqlutils contains functionality to wrap existing database functionality
+package sqlutils
 
 import (
 	"fmt"
@@ -9,10 +10,10 @@ import (
 // TransactionFunc defines the function signature needed to represent passing a transaction and returning a generic type
 type TransactionFunc[T any] func(*sqlx.Tx) (*T, error)
 
-// WithTransaction is a wrapper function which handles creating, committing or rollback a transaction
+// WithTransaction is a wrapper function which handles creating, committing or rolling back a transaction
 // If there are any errors when executing the txFunc, the tx is rolled back. Otherwise, the tx is committed.
-func WithTransaction[T any](s *Store, txFunc TransactionFunc[T]) (*T, error) {
-	tx, err := s.db.Beginx()
+func WithTransaction[T any](txPrep TransactionPreparer, txFunc TransactionFunc[T]) (*T, error) {
+	tx, err := txPrep.Beginx()
 	if err != nil {
 		return nil, fmt.Errorf("error creating transaction %w", err)
 	}
