+++ pkg/storage/store.go
@@ -35,6 +35,11 @@ func (s *Store) PrepareNamed(query string) (*sqlx.NamedStmt, error) {
 	return s.db.PrepareNamed(query)
 }
 
+// Beginx implements the TransactionPreparer interface
+func (s *Store) Beginx() (*sqlx.Tx, error) {
+	return s.db.Beginx()
+}
+
 // NewStore creates a new Store struct
 // The `db` property on the Store will always be a *sqlx.DB, but a notable difference in the DB is that if
 // config.UseIAM is true, that DB instance will be backed by a custom connector in iam_db.go that generates
