+++ pkg/storage/transactions_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/jmoiron/sqlx"
 
 	"github.com/cmsgov/easi-app/pkg/models"
+	"github.com/cmsgov/easi-app/pkg/sqlutils"
 )
 
 func (suite *StoreTestSuite) TestWithTransaction() {
@@ -14,7 +15,7 @@ func (suite *StoreTestSuite) TestWithTransaction() {
 	ctx := context.Background()
 	anonEua := "ANON"
 	suite.Run("No errors will commit a transaction", func() {
-		newTRB, err := WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
+		newTRB, err := sqlutils.WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
 
 			trb := models.NewTRBRequest(anonEua)
 			trb.Type = models.TRBTNeedHelp
@@ -36,7 +37,7 @@ func (suite *StoreTestSuite) TestWithTransaction() {
 	})
 
 	suite.Run("Errors will rollback a transaction", func() {
-		newTRB, err := WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
+		newTRB, err := sqlutils.WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
 
 			trb := models.NewTRBRequest(anonEua)
 			trb.Type = models.TRBTNeedHelp
@@ -55,7 +56,7 @@ func (suite *StoreTestSuite) TestWithTransaction() {
 
 	suite.Run("With Transaction can also perform discrete db actions not directly part of the transaction", func() {
 		var trbGlobal *models.TRBRequest
-		newTRB, err := WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
+		newTRB, err := sqlutils.WithTransaction[models.TRBRequest](suite.store, func(tx *sqlx.Tx) (*models.TRBRequest, error) {
 
 			trb := models.NewTRBRequest(anonEua)
 			trb.Type = models.TRBTNeedHelp
